Pylint: Add (commented out) missing-gettext plugin
The `missing-gettext` module is useful for identifying strings which
haven't been marked for gettext translation however it has no way to
distinguish between legitimate strings for gettext and strings used
within the code for other purposes (and logging which should normally be
in English so that the developers can understand) so can be annoying.
Because of this, give the option to uncomment it when needed.