Vim/Neovim: Warn about problem with YCM install
YouCompleteMe has an element which needs to be compiled. This is done by
entering the cloned repo and compiling the module. The module is only
cloned if it is supplied to Vundle through the vimrc, which only happens
if it is supported. This means that the Makefile rule for compiling
YouCompleteMe can fail if these certain conditions are met. As a result,
the Neovim Makefile rule will fail as it has this as a prerequisite.

Add a warning to the Makefile about this potential issue.